requirements.txt
mysql-connector-python==8.3.0
python-dotenv==0.19.0
SQLAlchemy==2.0.25
